Make truncations explicit
Bug: 129300035

+++ lib/fixed_point/include/lib/fixed_point.h
@@ -46,7 +46,7 @@ fp_32_64_div_32_32(struct fp_32_64 *result, uint32_t dividend, uint32_t divisor)
     tmp = ((uint64_t)dividend << 32) / divisor;
     rem = ((uint64_t)dividend << 32) % divisor;
     result->l0 = tmp >> 32;
-    result->l32 = tmp;
+    result->l32 = (uint32_t)tmp;
     tmp = ((uint64_t)rem << 32) / divisor;
     result->l64 = tmp;
 }
@@ -86,7 +86,7 @@ static uint32_t
 u32_mul_u64_fp32_64(uint64_t a, struct fp_32_64 b)
 {
     uint32_t a_r32 = a >> 32;
-    uint32_t a_0 = a;
+    uint32_t a_0 = (uint32_t)a;
     uint64_t res_l32;
     uint32_t ret;
 
@@ -96,7 +96,7 @@ u32_mul_u64_fp32_64(uint64_t a, struct fp_32_64 b)
     res_l32 += mul_u32_u32(a_0, b.l32, 0, -32);
     res_l32 += mul_u32_u32(a_r32, b.l64, 32, -64);
     res_l32 += mul_u32_u32(a_0, b.l64, 0, -64) >> 32; /* Improve rounding accuracy */
-    ret = (res_l32 >> 32) + ((uint32_t)res_l32 >> 31); /* Round to nearest integer */
+    ret = (uint32_t)((res_l32 >> 32) + ((uint32_t)res_l32 >> 31)); /* Round to nearest integer */
 
     debug_u32_mul_u64_fp32_64(a, b, res_l32, ret);
 
@@ -107,7 +107,7 @@ static uint64_t
 u64_mul_u64_fp32_64(uint64_t a, struct fp_32_64 b)
 {
     uint32_t a_r32 = a >> 32;
-    uint32_t a_0 = a;
+    uint32_t a_0 = (uint32_t)a;
     uint64_t res_0;
     uint64_t res_l32;
     uint32_t res_l32_32;
@@ -129,7 +129,7 @@ u64_mul_u64_fp32_64(uint64_t a, struct fp_32_64 b)
     tmp = mul_u32_u32(a_0, b.l64, 0, -64); /* Improve rounding accuracy */
     res_l32 += tmp >> 32;
     res_0 += res_l32 >> 32;
-    res_l32_32 = res_l32;
+    res_l32_32 = (uint32_t)res_l32;
     ret = res_0 +  (res_l32_32 >> 31); /* Round to nearest integer */
 
     debug_u64_mul_u64_fp32_64(a, b, res_0, res_l32_32, ret);
